+++ Source/Tests/Text/HCStringContainsTests.m
@@ -47,20 +47,6 @@ static NSString *EXCERPT = @"EXCERPT";
     assertDoesNotMatch(@"only part of excerpt", matcher, [EXCERPT substringFromIndex:1]);
 }
 
-- (void)testDeprecatedMatcher_EvaluatesToTrueIfArgumentContainsSpecifiedSubstring
-{
-    id <HCMatcher> deprecatedMatcher = containsString(EXCERPT);
-
-    assertMatches(@"excerpt at beginning", deprecatedMatcher, [EXCERPT stringByAppendingString:@"END"]);
-    assertMatches(@"excerpt at end", deprecatedMatcher, [@"START" stringByAppendingString:EXCERPT]);
-    assertMatches(@"excerpt in middle", deprecatedMatcher,
-            [[@"START" stringByAppendingString:EXCERPT] stringByAppendingString:@"END"]);
-    assertMatches(@"excerpt repeated", deprecatedMatcher, [EXCERPT stringByAppendingString:EXCERPT]);
-
-    assertDoesNotMatch(@"excerpt not in string", deprecatedMatcher, @"whatever");
-    assertDoesNotMatch(@"only part of excerpt", deprecatedMatcher, [EXCERPT substringFromIndex:1]);
-}
-
 - (void)testEvaluatesToTrueIfArgumentIsEqualToSubstring
 {
     assertMatches(@"excerpt is entire string", matcher, EXCERPT);